% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_amr_taxonomy.R
\name{generate_amr_taxonomy}
\alias{generate_amr_taxonomy}
\title{take raw AMR csv files to matrix}
\arguments{
\item{amr.count.table}{count of AMR genes with CVTERMID in the first row adn samples on columns}

\item{verbose}{true or false to keep all columns of AMR gene information or not}
}
\value{
data table of taxonomy of AMR genes in count table
}
\description{
take raw AMR csv files to matrix
}
\examples{
\dontrun{
generate_amr_taxonomy(amr.count.table = amr.count.table, verbose = FALSE)
}
}
