% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_amr_files.R
\name{read_in_amr_files}
\alias{read_in_amr_files}
\title{take raw AMR csv files to matrix}
\usage{
read_in_amr_files(path.to.files, coveragenumber)
}
\arguments{
\item{coveragenumber}{A number}

\item{path.tofile}{A file path}
}
\value{
matrix of AMR genes at a specific coverage
}
\description{
take raw AMR csv files to matrix
}
\examples{
read_in_amr_file("~/Desktop/my.files/", 80)
}
