% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amr_raw_to_phyloseq.R
\name{amr_raw_to_phyloseq}
\alias{amr_raw_to_phyloseq}
\title{Raw AMR files plus metadata to phyloseq object}
\arguments{
\item{x}{Data.}

\item{y}{More data.}

\item{amr.count.table}{count of AMR genes with CVTERMID in the first row and samples on columns}

\item{verbose}{true or false to keep all columns of AMR gene information or not}
}
\value{
data table of taxonomy of AMR genes in count table
}
\description{
given directory and metadata make phyloseq object \pkg{\link{phyloseq}} package required.
}
\examples{
\dontrun{
generate_amr_taxonomy(amr.count.table = amr.count.table, verbose = FALSE)
}
}
\seealso{
\pkg{\link{phyloseq}}
}
