% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_amr_files.R
\name{read_in_amr_files}
\alias{read_in_amr_files}
\title{take raw AMR csv files to matrix}
\arguments{
\item{path.to.amr.files}{A file path to AMR raw data csv}

\item{coveragenumber}{Minimum percentage of a gene that must
be covered 0 to 99}

\item{keepSNP}{true or false to keep AMR gene conferred by
one SNP change}
}
\value{
matrix of AMR genes at a specific coverage with or
without SNP associated
}
\description{
take raw AMR csv files to matrix
}
\examples{
\dontrun{
read_in_amr_files(path.to.amr.files = "~/Desktop/my.files/",
coveragenumber = 80, keepSNP = FALSE)
}
}
