% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_wimp_files.R
\name{read_in_wimp_files}
\alias{read_in_wimp_files}
\title{Convert raw WIMP CSV files to data table}
\usage{
read_in_wimp_files(path.to.wimp.files)
}
\arguments{
\item{path.to.wimp.files}{File path to folder containing raw WIMP files}
}
\value{
data.table of mb genes at a specific coverage
}
\description{
Convert raw WIMP CSV files to data table
}
\examples{
\dontrun{
read_in_wimp_files(path.to.wimp.files="~/Desktop/my.files")
}
}
