% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_wimp_files.R
\name{read_in_wimp_files}
\alias{read_in_wimp_files}
\title{take raw mb csv files to matrix}
\usage{
read_in_wimp_files(path.to.wimp.files)
}
\arguments{
\item{path.to.wimp.file}{A file path}
}
\value{
matrix of mb genes at a specific coverage
}
\description{
take raw mb csv files to matrix
}
\examples{
\dontrun{
read_in_mb_file(path.to.wimp.files="~/Desktop/my.files")
}
}
