% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wimp_raw_to_metagenomeseq.R
\name{wimp_raw_to_metagenomeseq}
\alias{wimp_raw_to_metagenomeseq}
\title{Raw WIMP files plus metadata to metagenomeseq object}
\usage{
wimp_raw_to_metagenomeseq(
  path.to.wimp.files,
  metadata,
  keep.unclassifed = FALSE,
  keep.human = FALSE
)
}
\arguments{
\item{path.to.wimp.files}{path to data of raw csv files from WIMP analysis}

\item{keep.human}{true or false to keep reads that classified to human default = FALSE}

\item{metdata}{dataframe of metadata with "filename" and "barcode"  columns required}

\item{keep.unclassified}{true or false to keep reads that do not classify below phylum, default = FALSE}
}
\value{
metagenomeseq object for downstream analysis with WIMP data
}
\description{
given wimp directory and metadata make phyloseq object \pkg{\link{metagenomeSeq}} package required.
}
\examples{
\dontrun{
wimp_raw_to_metagenomeseq(path.to.wimp.files= path/to/wimpfiles,
metadata = metadata, keep.unclassifed=FALSE, keep.human=FALSE)
}
}
\seealso{
\pkg{\link{metagenomeSeq}}
}
